Initial stages of the new Jerity Template System
Contains three new classes.  An abstract Template class which Chrome and
Content extend.  Chrome handles the general page layout that is
(usually) common to the entire site.  It also deals with resources for
the head of the document ensuring that there are no duplicates, etc.
The chrome can take one or more Content templates which are more
specific to the content actually being displayed on the page.