Add conditional proxying groundwork.
Support for conditional proxying must be added to each class that needs to use
it.  This is based on the implementation found in Propel ORM.

Currently there are some bugs that need to be addressed.  To keep things
simple, nested if statments are forbidden.  This only works properly when
_if(false) is called, however.  There is also a problem where when the initial
call is _if(true) and an _elseif() call is made, the _else() branch will be
incorrectly executed.

To solve the above problem it is quite likely that a state will need to be
maintained, 1. to inform whether an object has called _if(), and 2. to keep
track of whether a condition has evaluated to true.